api/apps/authentication/urls.py
from django.urls import path

from .views import RegistrationAPIView, HelloView
from rest_framework_simplejwt.views import TokenObtainPairView, TokenRefreshView

urlpatterns = [
    path('user/', RegistrationAPIView.as_view()),
    path('hello/', HelloView.as_view()),
    path('login/', TokenObtainPairView.as_view(), name='token_obtain_pair'),
    path('refresh/', TokenRefreshView.as_view(), name='token_refresh'),
]